Move the header content to its own file.
